requirements.txt
flask~=3.0.3
SQLAlchemy~=2.0.31
python-dotenv~=1.0.1